third_party/blink/web_tests/media/media-fragments/TC0015-expected.txt


Title: TC0015
Fragment: 't=,7'
Comment: Equivalent to #t=0,b. The media is requested from 0 to b.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
